src/app/pages/search/search.component.spec.ts
import { async, ComponentFixture, TestBed, tick } from '@angular/core/testing';
import { FormsModule } from '@angular/forms';
import { AngularFontAwesomeModule } from 'angular-font-awesome';

import { SearchComponent } from './search.component';

describe('SearchComponent', () => {
  let component: SearchComponent;
  let fixture: ComponentFixture<SearchComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      imports: [
        FormsModule,
        AngularFontAwesomeModule
      ],
      declarations: [ SearchComponent ],
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(SearchComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('verificando e inserindo no campo de pesquisa', async(() => {
    fixture.whenStable().then(() => {
      expect(component.value).toBe('');

      component.value = 'Carol';
      fixture.detectChanges();

      expect(component.value).toBe('Carol');

    });
  }));
});
